lesson-1: added App.kt, refactoring
